feat: AutoComplete SupportUpdate
