+++ src/BuildingBlocks/SearchEngine/Masa.BuildingBlocks.SearchEngine.AutoComplete/_Imports.cs
@@ -1,6 +1,6 @@
 // Copyright (c) MASA Stack All rights reserved.
 // Licensed under the MIT License. See LICENSE.txt in the project root for license information.
 
+global using Masa.BuildingBlocks.Data;
 global using Masa.BuildingBlocks.SearchEngine.AutoComplete.Options;
 global using Masa.BuildingBlocks.SearchEngine.AutoComplete.Response;
-
