+++ src/Contrib/SearchEngine/AutoComplete/Masa.Contrib.SearchEngine.AutoComplete.ElasticSearch/Options/AutoCompleteFactoryOptions.cs
@@ -1,0 +1,11 @@
+﻿// Copyright (c) MASA Stack All rights reserved.
+// Licensed under the MIT License. See LICENSE.txt in the project root for license information.
+
+// ReSharper disable once CheckNamespace
+
+namespace Masa.Contrib.SearchEngine.AutoComplete.ElasticSearch;
+
+public class AutoCompleteFactoryOptions: MasaFactoryOptions<AutoCompleteRelationsOptions>
+{
+
+}
