+++ src/Contrib/SearchEngine/AutoComplete/Masa.Contrib.SearchEngine.AutoComplete.ElasticSearch/Options/AutoCompleteRelationsOptions.cs
@@ -2,20 +2,21 @@
 // Licensed under the MIT License. See LICENSE.txt in the project root for license information.
 
 // ReSharper disable once CheckNamespace
+
 namespace Masa.Contrib.SearchEngine.AutoComplete.ElasticSearch;
 
-public class AutoCompleteRelationsOptions
-{
-    internal List<AutoCompleteRelations> Relations = new();
+// public class AutoCompleteRelationsOptions
+// {
+//     internal List<AutoCompleteRelations> Relations = new();
+//
+//     public AutoCompleteRelationsOptions AddRelation(AutoCompleteRelations options)
+//     {
+//         Relations.Add(options);
+//         return this;
+//     }
+// }
 
-    public AutoCompleteRelationsOptions AddRelation(AutoCompleteRelations options)
-    {
-        Relations.Add(options);
-        return this;
-    }
-}
-
-public class AutoCompleteRelations
+public class AutoCompleteRelationsOptions : MasaRelationOptions<IAutoCompleteClient>
 {
     internal bool IsDefault { get; }
 
@@ -29,27 +30,30 @@ public class AutoCompleteRelations
 
     internal bool EnableMultipleCondition { get; }
 
-    internal IElasticClient ElasticClient { get; }
+    internal string EsName { get; }
 
-    internal IMasaElasticClient MasaElasticClient { get; }
+    internal bool IsSupportUpdate { get; }
 
     internal SearchType DefaultSearchType { get; }
 
     internal string QueryAnalyzer { get; }
 
-    internal AutoCompleteRelations(
-        IElasticClient elasticClient,
-        IMasaElasticClient masaElasticClient,
+    internal AutoCompleteRelationsOptions(
+        string esName,
+        Func<IServiceProvider, IAutoCompleteClient> func,
+        bool isSupportUpdate,
         string indexName,
         string? alias,
         bool isDefault,
         Operator defaultOperator,
         SearchType defaultSearchType,
         bool enableMultipleCondition,
-        string queryAnalyzer)
+        string queryAnalyzer) : base(esName)
     {
-        ElasticClient = elasticClient;
-        MasaElasticClient = masaElasticClient;
+        EsName = esName;
+        Func = func;
+
+        IsSupportUpdate = isSupportUpdate;
         IndexName = indexName;
         Alias = alias;
         RealIndexName = alias ?? indexName;
