+++ src/Contrib/SearchEngine/AutoComplete/Masa.Contrib.SearchEngine.AutoComplete.ElasticSearch/_Imports.cs
@@ -13,7 +13,9 @@ global using Masa.Utils.Data.Elasticsearch.Options.Document.Delete;
 global using Masa.Utils.Data.Elasticsearch.Options.Document.Query;
 global using Masa.Utils.Data.Elasticsearch.Options.Document.Set;
 global using Masa.Utils.Data.Elasticsearch.Options.Index;
+global using Microsoft.Extensions.DependencyInjection;
 global using Microsoft.Extensions.DependencyInjection.Extensions;
 global using Microsoft.Extensions.Logging;
+global using Microsoft.Extensions.Options;
 global using Nest;
 global using DeleteResponse = Masa.BuildingBlocks.SearchEngine.AutoComplete.Response.DeleteResponse;
