+++ src/Utils/Data/Masa.Utils.Data.Elasticsearch/Options/MasaElasticsearchOptions.cs
@@ -1,0 +1,9 @@
+﻿// Copyright (c) MASA Stack All rights reserved.
+// Licensed under the MIT License. See LICENSE.txt in the project root for license information.
+
+namespace Masa.Utils.Data.Elasticsearch.Options;
+
+public class MasaElasticsearchOptions
+{
+    public Action<ElasticsearchOptions> Action { get; set; }
+}
