+++ test/Masa.Framework.IntegrationTests.EventBus/Infrastructure/CustomDbContext.cs
@@ -1,6 +1,8 @@
 ﻿// Copyright (c) MASA Stack All rights reserved.
 // Licensed under the MIT License. See LICENSE.txt in the project root for license information.
 
+using Masa.BuildingBlocks.Data;
+
 namespace Masa.Framework.IntegrationTests.EventBus.Infrastructure;
 
 public class CustomDbContext : MasaDbContext
